--- a/src/iter.rs
+++ b/src/iter.rs
@@ -1,5 +1,5 @@
 use core::slice;
-use std::convert::TryInto;
+use core::convert::TryInto;
 
 pub struct Bytes<'a> {
     slice: &'a [u8],