@AaronO 2022-08-27
I briefly played around with other perf improvements, and have so far achieved a ~20-25% improvement over `master` (benched on M1 Air, so without SIMD, though neon might be worth exploring).

```
## master
test req/req ... bench:                     241 ns/iter (+/- 2)
test req_short/req_short ... bench:         32 ns/iter (+/- 1)
test resp/resp ... bench:                   228 ns/iter (+/- 3)
test resp_short/resp_short ... bench:       40 ns/iter (+/- 0)

## PR #120 + other tweaks
test req/req ... bench:                     181 ns/iter (+/- 2)
test req_short/req_short ... bench:         25 ns/iter (+/- 4)
test resp/resp ... bench:                   183 ns/iter (+/- 6)
test resp_short/resp_short ... bench:       32 ns/iter (+/- 0)
```
(using `--output-format=bencher` for brevity)

I have a few other ideas for further improvements, happy to throw up PRs if you're interested