@bartlomieju 2022-08-29
@seanmonstar we'd love to try this change in our new server that uses `httparse`, any chance you could release a new version?