@bartlomieju 2022-08-29
Ah makes sense, @seanmonstar I'm fine waiting another few days for more changes like this!